src/requirements.txt
torchvision==0.6.0a0+82fd1c8
matplotlib==3.0.0
numpy==1.14.2
torch==1.5.0
auto_mix_prep==0.2.0
Pillow==7.2.0
